Added Photo and DisplayName properties
I'm not happy about this. Lync's GetContactInformation() call returns
syncronously, but isn't always ready immediatly, leading to us either
getting data in a state where it's not yet ready, or getting a
NotReadyException.

I'm mitigating this with some nast while loops and thread sleeps, but I
don't like it. Raising it as an issue to sort out later - there must be
a better way of doing this.